+++ NEWS.md
@@ -2,8 +2,12 @@
 
 ## BUG REPORTS
 
+- Fix classification bug due to repeated function name `call_component`.
+
 ## ENHANCEMENTS
 
+- Added data check and filter in `output_tally()` and `show_catalogue()`.
+- Enhanced `show_group_enrichment()` (#353).
 - Removed unnecessary CN classifications code in recent development.
 
 ## NEW FUNCTIONS